fix: response & add pre-next event button
