+++ src/App.css
@@ -12,7 +12,8 @@
 }
 @media screen and (max-width: 768px) {
     .grid-container {
-        margin-bottom:  50px;
+        margin-bottom: 30px;
+        flex-wrap: wrap;
     }
     
 }
@@ -24,6 +25,13 @@
     width: 5%;
     /* justify-content: center; */
     max-width: 100%;
+}
+
+@media screen and (max-width: 768px) {
+    .grid-col-1 {
+        width: 8%;
+        min-width: 60px;
+    }
 }
 
 .grid-col-1 div {
@@ -58,6 +66,13 @@
     padding: 0;
 }
 
+@media screen and (max-width: 768px) {
+    .grid-col-2 {
+        flex: 1;
+        min-width: 0;
+    }
+}
+
 .grid-col-2 img {
     width: 80%;
     height: 40%;
@@ -71,6 +86,13 @@
     display: flex;
     flex-direction: column;
     justify-content: center;
+}
+
+@media screen and (max-width: 768px) {
+    .grid-col-3 {
+        flex: 1;
+        min-width: 0;
+    }
 }
 
 .grid-col-3-title {
@@ -147,4 +169,40 @@
 .grid-col-4 div > p {
     margin: 0;
  
+}
+
+/* Additional mobile styles for CompanyInfoTable */
+@media screen and (max-width: 480px) {
+    .company-info-table {
+        font-size: 10px !important;
+    }
+    
+    .company-info-table td {
+        padding: 2px !important;
+        line-height: 1.2 !important;
+    }
+    
+    .company-info-table .label-cell {
+        font-size: 9px !important;
+        font-weight: bold;
+    }
+    
+    .company-info-table .data-cell {
+        font-size: 9px !important;
+    }
+}
+
+@media screen and (max-width: 360px) {
+    .company-info-table {
+        font-size: 8px !important;
+    }
+    
+    .company-info-table td {
+        padding: 1px !important;
+    }
+    
+    .company-info-table .label-cell,
+    .company-info-table .data-cell {
+        font-size: 8px !important;
+    }
 }
