+++ src/components/CountrySpecificComponent.jsx
@@ -34,18 +34,18 @@ const CountrySpecificComponent = ({ userCountry = 'vi' }) => {
   return (
     <div className="w-full h-full flex flex-col">
       {/* Global Logo - Planet Earth */}
-      <div className="flex-1 flex items-center justify-center p-4">
+      <div className="flex-1 flex items-center justify-center p-2 sm:p-4">
         <div className="">
           <img
             src={currentCountry.flag}
             alt="Country"
-            className="max-w-[200px] max-h-[200px] object-contain mx-auto"
+            className="w-full h-full max-w-[100px] max-h-[100px] sm:max-w-[150px] sm:max-h-[150px] md:max-w-[200px] md:max-h-[200px] object-contain mx-auto"
           />
         </div>
       </div>
       
       {/* Global Website Information */}
-      <div className="p-4 text-center">
+      <div className="p-2 sm:p-4 text-center">
       </div>
     </div>
   );
